+++ src/i18n/languages/fiFI.ts
@@ -188,6 +188,7 @@ const dictionary: RawTranslationPackage = {
     SEC: 'SEK',
     SECH: 'SEKH',
     SECOND: 'SEKUNNIT',
+    SEQUENCE: 'JAKSO',
     SHEET: 'TAULUKKO',
     SHEETS: 'TAULUKOT',
     SIN: 'SIN',