+++ src/i18n/languages/huHU.ts
@@ -188,6 +188,7 @@ const dictionary: RawTranslationPackage = {
     SEC: 'SEC',
     SECH: 'SECH',
     SECOND: 'MPERC',
+    SEQUENCE: 'SOROZAT',
     SHEET: 'LAP',
     SHEETS: 'LAPOK',
     SIN: 'SIN',